Changed handling of axes when projecting onto EOFs
Problems could occur when a time dimension was present but was not
detected by cdms2. In this case a confusing error about missing
values would be produced.

The handling ox axes is now based on inspection of array shapes,
avoiding the need for the user to specify if a time dimension is
present in the numpy interface. The alleviates the problems seen
in the cdms2 interface also.